# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Integration service Arquivei",
        "version": "v1"
    },
    "paths": {
        "/api/nfe/{access_key}": {
            "get": {
                "tags": [
                    "NFe"
                ],
                "summary": "Return NFe",
                "description": "Return a object of NFe",
                "operationId": "Nfe/show",
                "parameters": [
                    {
                        "name": "access_key",
                        "in": "path",
                        "description": "Access key for NFe",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Object of NFe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NFe"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NFe not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "x-api-id": []
                    },
                    {
                        "x-api-key": []
                    }
                ]
            }
        },
        "/api/download/{access_key}": {
            "get": {
                "tags": [
                    "NFe"
                ],
                "summary": "Download XML NFe",
                "description": "Download XML of NFe",
                "operationId": "Nfe/download",
                "parameters": [
                    {
                        "name": "access_key",
                        "in": "path",
                        "description": "Access key for NFe",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stream file xml with named is access key",
                        "content": {
                            "application/octet-stream": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NFe not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "x-api-id": []
                    },
                    {
                        "x-api-key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Error": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "NFe": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "access_key": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "xml": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "x-api-id": {
                "type": "apiKey",
                "description": "Authorization API ID",
                "name": "x-api-id",
                "in": "header"
            },
            "x-api-key": {
                "type": "apiKey",
                "description": "Authorization API KEY",
                "name": "x-api-key",
                "in": "header"
            }
        }
    }
}